# api_docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Online Shop API",
    "description": "Android ilova uchun online do'kon backend API. Savat, buyurtma va mahsulotlar bilan ishlash uchun.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/customers": {
      "post": {
        "tags": [
          "customers"
        ],
        "summary": "Yangi mijozni ro'yxatdan o'tkazish",
        "description": "Full name, phone va optional default_address asosida yangi customer yaratadi.",
        "operationId": "create_customer_api_v1_customers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/customers/login": {
      "post": {
        "tags": [
          "customers"
        ],
        "summary": "Mijoz login qilishi",
        "description": "Telefon raqami va parol orqali login qilish.",
        "operationId": "login_customer_api_v1_customers_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/customers/{customer_id}": {
      "get": {
        "tags": [
          "customers"
        ],
        "summary": "Customer ma'lumotlarini olish",
        "operationId": "get_customer_api_v1_customers__customer_id__get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Customer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/customers/by-phone": {
      "get": {
        "tags": [
          "customers"
        ],
        "summary": "Telefon raqam bo'yicha customer topish",
        "operationId": "get_customer_by_phone_api_v1_customers_by_phone_get",
        "parameters": [
          {
            "name": "phone",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Phone"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomerCreate": {
        "properties": {
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "default_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Address"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "phone",
          "password"
        ],
        "title": "CustomerCreate"
      },
      "CustomerLogin": {
        "properties": {
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "password"
        ],
        "title": "CustomerLogin"
      },
      "CustomerOut": {
        "properties": {
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "default_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Address"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "phone",
          "id"
        ],
        "title": "CustomerOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}